+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 
   - Modify fields on connect method.
   - Support setting custom root_helper.
+  - Setting default project_id and user_id.
 
 - Bug fixes: